now we have a dedicated file for the debugger and command handling
