+++ .gitignore
@@ -1,6 +1,6 @@
 build
 build/*
-dbg
+wg
 disass
 elf
